05_25Merge

+++ Capstone_2022_Japgokbap/Assets/Scripts/PlayerController.cs
@@ -42,8 +42,8 @@ public class PlayerController : MonoBehaviour
     [HideInInspector] public List<int> PlayerSkill = new List<int>();
     
     [Header("PlayerStatus")]
-    public int playerMaxHP;
-    public int playerCurrentHP;
+    public float playerMaxHP;
+    public float playerCurrentHP;
     public int playerOffensePower;
     public int playerDeffencePower;
     public float playerMoveSpeed;
@@ -246,6 +246,7 @@ public class PlayerController : MonoBehaviour
                 Destroy(other.gameObject);
             break;
             case "Potion":
+                //potion의 고유 hp로 바꿔야함.
                 playerCurrentHP += 10;
                 Debug.Log("CurrentHP has recovered");
                 if(playerCurrentHP >= playerMaxHP){
